refactor(storage): move mapping impl in state instead of msg
